Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src-attr-elem/script-src-elem-blocked-attr-allowed-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'script-src-elem'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'script-src-attr'.



Harness Error (TIMEOUT), message = null

NOTRUN Should fire a security policy violation for the attribute
PASS Should execute the inline script attribute
